Grow image with standard blue fill.
The previous behavior was to extend the edges of the existing image.
However, firefox trims screenshots to only include the top active part
of the document rather than screenshoting the viewport so this means
that things at the bottom edge get extending rather than the empty
color. So just force the image growth to use the standard emtpy fill
color (#1289af). The extend algorithm in probably more ideal if
firefox begins screenshotting properly because it would account for
the body background color being changed by the test case.