Grammar: add spaces, join strings, reduce noise.
Previously mandatory and optional spaces were not being handled
correctly such that most multi-word CSS property values were being
concatenated. In reality many of these optional spaces aren't
optional but it's really in issue during generation. So instead of
stripping we just make optional spaces a single space string.

This balloons the generators a bit so to reduce the size (even beyond
the starting point) we join all strings next to each other in a :cat
(juxtaposition) into single concatenated strings.

Expand the grammar mangling and replace the items that were causing
unicode noise to be generated in the pages. Also, replace a number of
the regex values being used to generate various numbers and other
things with some native Clojure generators this should be more
efficient and shrink better.

Add capability to mend to output the intermediate grammar tree as EDN.