More efficient app. Add firefox/chrome data.
This version of the app splits the flat view into 100 items at a time
and only create the TAPV popup data when it's actually requested by
the user instead of as hidden data in every table cell.

Also, switch top-level data files from EDN to transit. Transit is
a bit more efficient for loading and parsing in the browser than EDN.